OPENNLP-970: Adds confusion matrix to NameFinder
Adds a refined report to NameFinder
